Refactor form out of AddPatient component
